Enhance configuration and logging for trading operations
- Added DHAN_AUTO_RESTART_ON_RENEW option in .env.example
- Updated EOD squareoff time to 15:15 IST in multiple files
- Improved logging for order fill polling and market exit operations
- Adjusted market position reconciliation to reset daily P&L only before market opens
- Refined VWAP calculation in indicators.py for better accuracy
- Removed deprecated migration script for enhanced journal fields

+++ yukti/data/migrations/versions/005_journal_reflection_fields.py
@@ -18,7 +18,7 @@ depends_on = None
 def upgrade() -> None:
     op.add_column(
         "journal_entries",
-        sa.Column("quality_score", sa.Integer(), nullable=True),
+        sa.Column("quality_score", sa.Float(), nullable=True),
     )
     op.add_column(
         "journal_entries",
